fix(java): Line up protobuf type handling with our other generators (#6322)
* fix(java): Line up protobuf type handling with our other generators

* Actually revert back to using long for uint64

* run fixtures

* Add changelog yml

* chore: update changelog

---------

+++ generators/java/generator-utils/src/main/java/com/fern/java/generators/AliasGenerator.java
@@ -293,7 +293,7 @@ public final class AliasGenerator extends AbstractTypeGenerator {
 
         @Override
         public CodeBlock visitUint() {
-            return CodeBlock.of("return $T.$L($L)", Long.class, "toString", VALUE_FIELD_NAME);
+            return CodeBlock.of("return $T.$L($L)", Integer.class, "toString", VALUE_FIELD_NAME);
         }
 
         @Override
